+++ tests/test_thermochem.py
@@ -1,6 +1,7 @@
 import os
 import pytest
 import numpy as np
+
 from autode import Molecule, Atom, Calculation, HessianKeywords
 from autode.transition_states import TSguess, TransitionState
 from autode.thermochemistry import calculate_thermo_cont
@@ -9,7 +10,7 @@ from autode.values import Energy
 from autode.species import Species
 from autode.methods import ORCA, G09
 from . import testutils
-from autode.thermochemistry.igm import _q_rot_igm, _s_rot_rr, _entropy, _zpe
+from autode.thermochemistry.igm import _q_rot_igm, _s_rot_rr, _zpe
 
 here = os.path.dirname(os.path.abspath(__file__))
 
@@ -250,10 +251,10 @@ def test_unknown_entropy_method():
 
     h2 = Molecule(atoms=[Atom("H"), Atom("H", x=0.7)])
 
-    with pytest.raises(NotImplementedError):
-        _ = _entropy(
+    with pytest.raises(KeyError):
+        _ = calculate_thermo_cont(
             species=h2,
-            method="an_unkown_method",
+            lfm_method="an_unkown_method",
             temp=298,
             ss="1M",
             shift=100,
@@ -299,3 +300,22 @@ def test_calc_thermo_with_calc():
 
     assert os.path.exists("h2_calc_hess_orca.inp")
     assert "B3LYP " in open("h2_calc_hess_orca.inp", "r").readline()
+
+
+@testutils.work_in_zipped_dir(os.path.join(here, "data", "thermochem.zip"))
+def test_themochem_minenkov_is_close_to_grimme():
+
+    mol = Molecule("alkane_hess_orca.xyz")
+    calc = Calculation(
+        name="tmp", molecule=mol, method=orca, keywords=orca.keywords.hess
+    )
+    calc.set_output_filename("alkane_hess_orca.hess")
+
+    mol.calc_thermo(calc=calc, ss="1atm", sn=1, lfm_method="grimme")
+    G_grimme = mol.g_cont
+
+    mol.calc_thermo(calc=calc, ss="1atm", sn=1, lfm_method="minenkov")
+    G_minenkov = mol.g_cont
+
+    mad = abs(float(G_grimme.to("kcal mol-1") - G_minenkov.to("kcal mol-1")))
+    assert 1e-5 < mad < 1  # Absolute differnce should be small but non zero